21-10-22 Brain_scene_update
- rfid 전달 방식 변경, address.cs 사용
- cam1만 사용, brain을 rotate
- mic input update
