--- scienscope_2021/Assets/Mirza Beig/Particle Systems/Ultimate VFX/Demos/Starfall/Starfall.unitypackage.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7e6f21b8f25df3347b5123c3e48ec631
-timeCreated: 1498652894
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
