--- scienscope_2021/Assets/Mirza Beig/Particle Systems/Ultimate VFX/Demos/The Last Particle/The Last Particle.unitypackage.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ccadd00ebe285674fb6c60a4399203a8
-timeCreated: 1497588141
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
