+++ scienscope_2021/Assets/Scenes/2_brain_Profiles/brain_scene_control.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9ae389b95b0c844d8755b543a7248f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
