+++ scienscope_2021/ProjectSettings/EditorBuildSettings.asset
@@ -11,7 +11,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/1_RFID_waiting.unity
     guid: ac6878f61ba08a946a6ff7e8086f8c98
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/2_brain.unity
     guid: 88d36f03dba9045c786977210a7f508f
   - enabled: 0
@@ -26,7 +26,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/template_scene.unity
     guid: 1f013b3ee419c8241ae6bd40b0dfe6ac
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/vacuum_cleaner/vc_test_scene.unity
     guid: 0d607ca66882a0940b8472982e138361
   - enabled: 0
