+++ scienscope_2021/UserSettings/EditorUserSettings.asset
@@ -6,51 +6,13 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c0f2f050319383866333e243d04
+      value: 22424703114646680e0b0227036c491115020d27132b2a352c271821adf63707f6e225fdd8293c393707a1290a371526
       flags: 0
     RecentlyUsedScenePath-1:
-      value: 224247031146467d3822201d3816723138283a180d01087f0b3c113fddc22639ebe909cde23730722c0ce6281d
-      flags: 0
-    RecentlyUsedScenePath-2:
-      value: 224247031146467d3822201d3816723138283a180d01087f0b3c113fddc22639ebe909d9e6282b2f7717e1351027
-      flags: 0
-    RecentlyUsedScenePath-3:
-      value: 224247031146464f08031c2e11375a2f05141d242966333e243d04
-      flags: 0
-    RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b0227036c0c2f0612173a202d68252320092a
-      flags: 0
-    RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c0d2f140519232266333e243d04
-      flags: 0
-    RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c4d161f132738292922353f165326ece92021
-      flags: 0
-    RecentlyUsedScenePath-7:
-<<<<<<< HEAD
-      value: 22424703114646680e0b0227036c4d161f132738292922353f165326ece92021
-=======
-      value: 2242470311464676041c162350015a191158282b3e3c2f33212c5d00fbf3203deff479dceb2e36313816ea7c32183970c1090c05f318455a5143bd3b1deb191e10d11d852cee275c833d19c3c2549ff6dbdfc0f3
->>>>>>> main
-      flags: 0
-    RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c0e2f2431310e133f273939201334acf53a31f6fe
       flags: 0
-    RecentlyUsedScenePath-9:
-<<<<<<< HEAD
+    RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c0d2f140519232266333e243d04
-=======
-      value: 22424703114646680e0b0227036c491115020d27132b2a352c271821adf63707f6e225fdd8293c393707a1290a371526
->>>>>>> main
-      flags: 0
-    UnityEditor.ShaderGraph.Blackboard:
-      value: 18135939215a0a5004000b0e15254b524c030a3f2964643d120d1230e9e93a3fd6e826abbd2e2d293c4ead313b08042de6030a0afa240c0d020be94c4ba75e435d8715fa32c70d15d11612dacc11fee5d3c5d1fe9ab1b0968786f9abada9bbaaa6a0f3a9f2fcae8ce8a2c182a6d8c2869bd1f5dc8e9c9b97ea9c819ffcca90898d86d598899ac5
-      flags: 0
-    UnityEditor.ShaderGraph.FloatingWindowsLayout2:
-      value: 181344140043005e1a220d3b1f364b524c0c5a27130c293326201334cee5322ca0bd30e8eb293a707b0fd0180b3d0a36fc0d3d04e649500d1002ee0b5dbd1d2c27c00ad113cb1e10e41f1addc80993b98d9884a69ae6d8f0d1cda9e8fbfefaf9f9dea3fdb9ade882f0f7b0e1e380cafbf2c3adc18e9cd285a2908b82ec869c808a85d6878bde80838bd5f1dff7dce5bdd1fbf5c4c1ebcffdcbc8c4e530e2fe37373f3fefa1a5ab
-      flags: 0
-    UnityEditor.ShaderGraph.InspectorWindow:
-      value: 18135939215a0a5004000b0e15254b524c1119263f2d6a722016393ce1eb3d36e5d339f9a5602b2e2c07a37e0901373ae01e0008f707250d171df81a53a5485d41895ac825e0100ec20313c0d91cddccd3d0c7efcca9bd80908fecb0f9cfddf1eff4e7a1b1eae482f0fcaee1e1928b86d888ed969b928791ec84878be98383828e87dd818e9ac5
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
